+++ src/main/java/com/apex/tool/colorization/ColorData.java
@@ -3,8 +3,8 @@ package com.apex.tool.colorization;
 import com.apex.core.Constants;
 
 public class ColorData {
-    public float[] barycentric;
+    public double[] barycentric;
     public float u0, v0, u1, v1, u2, v2;
-    public float lightFactor = 1;
-    public float MIN_LIGHT_FACTOR = Constants.MIN_LIGHT_FACTOR;
+    public double lightFactor = 1;
+    public double MIN_LIGHT_FACTOR = Constants.MIN_LIGHT_FACTOR;
 }
